requirements.txt
pywebview
bcrypt
google-generativeai
flask
flask-cors
python-dotenv
SpeechRecognition
pyaudio
edge-tts
pygame
